Added updated Unit tests for classes Contract and ContractManager.
All tests covering 100% of all classes and methods and 90-100% of lines. For better understanding also added comments. The language of tests is english as we discussed and decided earlier.

There is no tests for logs and console output (System.out.println) because I do not know if we really need that in our project.